Só falta revisar o código do file upload

+++ src/gwt/material/design/components/client/ui/misc/fileUpload/js/JsProgressData.java
@@ -32,10 +32,10 @@ import jsinterop.annotations.JsType;
 public class JsProgressData {
 
 	@JsProperty
-    public long loaded;
+    public int loaded;
 	
 	@JsProperty
-    public long total;
+    public int total;
 	
 	@JsProperty
     public JsFile[] files;
